requirements.txt
requests
pandas
python-dotenv
pyarrow
psycopg2-binary
sqlalchemy
